requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
confluent-kafka==1.9.2
idna==3.4
requests==2.28.2
urllib3==1.26.14
PyYAML~=6.0